💄[Design] CardSkeleton 애니메이션 추가, 색상 대비 높임
