+++ src/components/commons/card/CardSkeleton/index.ts
@@ -1,0 +1,1 @@
+export { default as CardSkeleton } from "./CardSkeleton";
